Move Navbar to shared folder
