+++ app/screens/layout/layout.less
@@ -1,5 +1,3 @@
-@import './navbar';
-
 /**
  * Code based on https://github.com/philipwalton/solved-by-flexbox/blob/master/assets/css/components/site.css
