+++ app/shared/shared.less
@@ -5,5 +5,6 @@
 @import './footer/footer';
 @import './form-fields/form-fields';
 @import './modals/modals';
+@import './navbar/navbar';
 @import './resource-icons/resource-icons';
 @import './resource-list/resource-list';
